Rename files to make them eligible Python modules.
This also ensures that all Python files are named using snake_case.
